# nodejs_space/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/upload/presigned": {
      "post": {
        "operationId": "UploadController_getPresignedUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUploadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Upload"
        ]
      }
    },
    "/api/upload/complete": {
      "post": {
        "operationId": "UploadController_completeUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Upload"
        ]
      }
    },
    "/api/files/{id}/url": {
      "get": {
        "operationId": "UploadController_getFileUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUrlResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Upload"
        ]
      }
    },
    "/api/files/{id}": {
      "delete": {
        "operationId": "UploadController_deleteFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Upload"
        ]
      }
    },
...
    "/api/terms": {
      "get": {
        "operationId": "TermsController_findAll",
        "parameters": [
          {
            "name": "academicYearId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TermListItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Terms"
        ]
      },
      "post": {
        "operationId": "TermsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTermDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Terms"
        ]
      }
    },
    "/api/terms/{id}": {
      "delete": {
        "operationId": "TermsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Terms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Api",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PresignedUploadDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "fileName",
          "contentType"
        ]
      },
      "PresignedUploadResponseDto": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "cloud_storage_path": {
            "type": "string"
          }
        },
        "required": [
          "uploadUrl",
          "cloud_storage_path"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {
          "cloud_storage_path": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "fileSize": {
            "type": "number"
          }
        },
        "required": [
          "cloud_storage_path",
          "fileName",
          "contentType"
        ]
      },
      "CompleteUploadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "cloud_storage_path": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "cloud_storage_path",
          "fileName"
        ]
      },
      "FileUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "TermListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "academicYearId": {
            "type": "string"
          },
          "academicYearName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "academicYearId",
          "academicYearName"
        ]
      },
      "CreateTermDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "academicYearId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "name",
          "startDate",
          "endDate",
          "academicYearId"
        ]
      },
      "TermResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "academicYearId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "academicYearId"
        ]
      },
...
    }
  }
}